Changed to get the link in bio
Changed "you may dm us to get the link" to "get  the link in our bio".
an output_link.txt file is added where the link of that story will written here. The person who post just need to copy the link in this output_link.txt file to get the link, instead of look into the excel sheet.